#134 refactor: FE 공부방 페이지 - 커스텀훅을 통한 비즈니스 로직 분리
